fix(tourism): merge

+++ frontend/core-ui/src/modules/settings/structure/components/branches/BranchColumns.tsx
@@ -4,13 +4,13 @@ import {
   Button,
   Input,
   RecordTable,
-  RecordTableCellContent,
-  RecordTableCellDisplay,
-  RecordTableCellTrigger,
-  RecordTablePopover,
+  RecordTableInlineCell,
+  Popover,
   RecordTableTree,
+  Spinner,
   Textarea,
   TextOverflowTooltip,
+  useConfirm,
   useQueryState,
 } from 'erxes-ui';
 import { useSetAtom } from 'jotai';
@@ -70,14 +70,21 @@ export const BranchRemoveCell = ({
 }: {
   cell: Cell<IBranchListItem, unknown>;
 }) => {
-  const { _id } = cell.row.original;
+  const { confirm } = useConfirm();
+  const { _id, title } = cell.row.original;
   const { handleRemove, loading } = useRemoveBranch();
+
   const onRemove = () => {
-    handleRemove({
-      variables: {
-        ids: [_id],
-      },
-    });
+    confirm({
+      message: `Are you sure you want to remove '${title}'`,
+      options: { confirmationValue: 'delete' },
+    }).then(() =>
+      handleRemove({
+        variables: {
+          ids: [_id],
+        },
+      }),
+    );
   };
   return (
     <Button
@@ -86,61 +93,13 @@ export const BranchRemoveCell = ({
       onClick={onRemove}
       className="text-destructive bg-destructive/10"
     >
-      <IconTrash size={12} />
+      {loading ? <Spinner /> : <IconTrash size={12} />}
     </Button>
   );
 };
 
 export const BranchColumns: ColumnDef<IBranchListItem>[] = [
   RecordTable.checkboxColumn as ColumnDef<IBranchListItem>,
-  {
-    id: 'title',
-    accessorKey: 'title',
-    header: () => <RecordTable.InlineHead label="title" />,
-    cell: ({ cell }) => {
-      const { title, _id } = cell.row.original;
-      const [_title, setTitle] = useState<string>(title);
-      const { branchesEdit, loading } = useBranchInlineEdit();
-      const [open, setOpen] = useState<boolean>(false);
-
-      const onSave = () => {
-        if (_title !== title) {
-          branchesEdit({ variables: { id: _id, title: _title } }, ['title']);
-        }
-      };
-
-      const onChange = (el: ChangeEvent<HTMLInputElement>) => {
-        const { value } = el.currentTarget;
-        setTitle(value);
-      };
-
-      return (
-        <RecordTablePopover
-          open={open}
-          onOpenChange={(open) => {
-            setOpen(open);
-            if (!open) {
-              onSave();
-            }
-          }}
-        >
-          <RecordTableCellTrigger>
-            <RecordTableTree.Trigger
-              order={cell.row.original.order}
-              name={cell.getValue() as string}
-              hasChildren={cell.row.original.hasChildren as boolean}
-            >
-              <TextOverflowTooltip value={cell.getValue() as string} />
-            </RecordTableTree.Trigger>
-          </RecordTableCellTrigger>
-          <RecordTableCellContent>
-            <Input value={_title} onChange={onChange} />
-          </RecordTableCellContent>
-        </RecordTablePopover>
-      );
-    },
-    size: 250,
-  },
   {
     id: 'code',
     accessorKey: 'code',
@@ -158,43 +117,96 @@ export const BranchColumns: ColumnDef<IBranchListItem>[] = [
       };
 
       const onChange = (el: ChangeEvent<HTMLInputElement>) => {
-        const { value } = el.currentTarget;
-        setCode(value);
+        setCode(el.currentTarget.value);
       };
 
       return (
-        <RecordTablePopover
+        <Popover
           open={open}
           onOpenChange={(open) => {
             setOpen(open);
-            if (!open) {
-              onSave();
-            }
+            if (!open) onSave();
           }}
         >
-          <RecordTableCellTrigger>
-            {cell.getValue() as string}
-          </RecordTableCellTrigger>
-          <RecordTableCellContent>
-            <Input value={_code} onChange={onChange} />
-          </RecordTableCellContent>
-        </RecordTablePopover>
+          <RecordTableInlineCell.Trigger>
+            <RecordTableTree.Trigger
+              order={cell.row.original.order}
+              name={cell.getValue() as string}
+              hasChildren={cell.row.original.hasChildren as boolean}
+            >
+              <TextOverflowTooltip value={cell.getValue() as string} />
+            </RecordTableTree.Trigger>
+          </RecordTableInlineCell.Trigger>
+          <RecordTableInlineCell.Content>
+            <Input value={_code} onChange={onChange} disabled={loading} />
+          </RecordTableInlineCell.Content>
+        </Popover>
       );
     },
+  },
+  {
+    id: 'title',
+    accessorKey: 'title',
+    header: () => <RecordTable.InlineHead label="title" />,
+    cell: ({ cell }) => {
+      const { title, _id, code } = cell.row.original;
+      const [_title, setTitle] = useState<string>(title);
+      const { branchesEdit, loading } = useBranchInlineEdit();
+      const [open, setOpen] = useState<boolean>(false);
+
+      const onSave = () => {
+        if (_title !== title) {
+          branchesEdit({ variables: { id: _id, title: _title, code } }, [
+            'title',
+            'code',
+          ]);
+        }
+      };
+
+      const onChange = (el: ChangeEvent<HTMLInputElement>) => {
+        setTitle(el.currentTarget.value);
+      };
+
+      return (
+        <Popover
+          open={open}
+          onOpenChange={(open) => {
+            setOpen(open);
+            if (!open) onSave();
+          }}
+        >
+          <RecordTableInlineCell.Trigger>
+            {cell.getValue() as string}
+          </RecordTableInlineCell.Trigger>
+          <RecordTableInlineCell.Content>
+            <Input value={_title} onChange={onChange} disabled={loading} />
+          </RecordTableInlineCell.Content>
+        </Popover>
+      );
+    },
+    size: 250,
   },
   {
     id: 'parentId',
     accessorKey: 'parentId',
     header: () => <RecordTable.InlineHead label="parent" />,
     cell: ({ cell }) => {
+      const { parentId, _id, code } = cell.row.original;
+      const { branchesEdit, loading } = useBranchInlineEdit();
+
       return (
-        <div>
-          <SelectBranch
-            className="shadow-none bg-transparent"
-            value={cell.getValue() as string}
-            onValueChange={() => {}}
-          />
-        </div>
+        <SelectBranches.InlineCell
+          mode="single"
+          value={cell.getValue() as string}
+          onValueChange={(value) => {
+            if (value !== parentId) {
+              branchesEdit(
+                { variables: { id: _id, parentId: value, code: code } },
+                ['parentId', 'code'],
+              );
+            }
+          }}
+        />
       );
     },
   },
@@ -203,16 +215,17 @@ export const BranchColumns: ColumnDef<IBranchListItem>[] = [
     accessorKey: 'address',
     header: () => <RecordTable.InlineHead label="address" />,
     cell: ({ cell }) => {
-      const { address, _id } = cell.row.original;
+      const { address, _id, code } = cell.row.original;
       const [_address, setAddress] = useState<string>(address);
       const { branchesEdit, loading } = useBranchInlineEdit();
       const [open, setOpen] = useState<boolean>(false);
 
       const onSave = () => {
         if (_address !== address) {
-          branchesEdit({ variables: { id: _id, address: _address } }, [
-            'address',
-          ]);
+          branchesEdit(
+            { variables: { id: _id, address: _address, code: code } },
+            ['address', 'code'],
+          );
         }
       };
 
@@ -221,7 +234,7 @@ export const BranchColumns: ColumnDef<IBranchListItem>[] = [
         setAddress(value);
       };
       return (
-        <RecordTablePopover
+        <Popover
           open={open}
           onOpenChange={(open) => {
             setOpen(open);
@@ -230,13 +243,13 @@ export const BranchColumns: ColumnDef<IBranchListItem>[] = [
             }
           }}
         >
-          <RecordTableCellTrigger>
+          <RecordTableInlineCell.Trigger>
             <TextOverflowTooltip value={cell.getValue() as string} />
-          </RecordTableCellTrigger>
-          <RecordTableCellContent>
+          </RecordTableInlineCell.Trigger>
+          <RecordTableInlineCell.Content>
             <Textarea value={_address} onChange={onChange} />
-          </RecordTableCellContent>
-        </RecordTablePopover>
+          </RecordTableInlineCell.Content>
+        </Popover>
       );
     },
     size: 250,
@@ -247,9 +260,9 @@ export const BranchColumns: ColumnDef<IBranchListItem>[] = [
     header: () => <RecordTable.InlineHead label="team member count" />,
     cell: ({ cell }) => {
       return (
-        <RecordTableCellDisplay className="text-center flex w-full justify-center">
+        <RecordTableInlineCell className="text-center flex w-full justify-center">
           <Badge variant={'secondary'}>{cell.getValue() as number}</Badge>
-        </RecordTableCellDisplay>
+        </RecordTableInlineCell>
       );
     },
   },
@@ -258,11 +271,11 @@ export const BranchColumns: ColumnDef<IBranchListItem>[] = [
     header: () => <RecordTable.InlineHead label="Actions" />,
     cell: ({ cell }) => {
       return (
-        <RecordTableCellDisplay className="flex justify-center gap-1 [&>button]:px-2">
+        <RecordTableInlineCell className="flex justify-center gap-1 [&>button]:px-2">
           <BranchWorkingHoursColumnCell cell={cell} />
           <BranchEditColumnCell cell={cell} />
           <BranchRemoveCell cell={cell} />
-        </RecordTableCellDisplay>
+        </RecordTableInlineCell>
       );
     },
   },
