+++ frontend/core-ui/src/modules/settings/structure/components/departments/DepartmentColumns.tsx
@@ -5,19 +5,23 @@ import {
   Button,
   Input,
   RecordTable,
-  RecordTableCellContent,
-  RecordTableCellDisplay,
-  RecordTableCellTrigger,
-  RecordTablePopover,
+  RecordTableInlineCell,
+  Popover,
   RecordTableTree,
+  Spinner,
   TextOverflowTooltip,
+  useConfirm,
   useQueryState,
 } from 'erxes-ui';
 import { IDepartmentListItem } from '../../types/department';
 import { useSetAtom } from 'jotai';
 import { renderingDepartmentDetailAtom } from '../../states/renderingDepartmentDetail';
-import { AssignMember, AssignMemberTrigger } from 'ui-modules';
-import { useRemoveDepartment } from '../../hooks/useDepartmentActions';
+import { SelectMember } from 'ui-modules';
+import {
+  useDepartmentInlineEdit,
+  useRemoveDepartment,
+} from '../../hooks/useDepartmentActions';
+import { useState } from 'react';
 
 export const DepartmentWorkingHoursColumnCell = ({
   cell,
@@ -70,14 +74,20 @@ export const DepartmentRemoveCell = ({
 }: {
   cell: Cell<IDepartmentListItem, unknown>;
 }) => {
-  const { _id } = cell.row.original;
+  const { _id, title } = cell.row.original;
+  const { confirm } = useConfirm();
   const { handleRemove, loading } = useRemoveDepartment();
   const onRemove = () => {
-    handleRemove({
-      variables: {
-        ids: [_id],
-      },
-    });
+    confirm({
+      message: `Are you sure you want to remove '${title}'`,
+      options: { confirmationValue: 'delete' },
+    }).then(() =>
+      handleRemove({
+        variables: {
+          ids: [_id],
+        },
+      }),
+    );
   };
   return (
     <Button
@@ -86,7 +96,7 @@ export const DepartmentRemoveCell = ({
       onClick={onRemove}
       className="text-destructive bg-destructive/10"
     >
-      <IconTrash size={12} />
+      {loading ? <Spinner /> : <IconTrash size={12} />}
     </Button>
   );
 };
@@ -98,15 +108,42 @@ export const DepartmentColumns: ColumnDef<IDepartmentListItem>[] = [
     accessorKey: 'code',
     header: () => <RecordTable.InlineHead icon={IconHash} label="code" />,
     cell: ({ cell }) => {
+      const { departmentsEdit, loading } = useDepartmentInlineEdit();
+      const { _id, code } = cell.row.original;
+      const [_code, setCode] = useState<string>(code);
+      const [open, setOpen] = useState<boolean>(false);
+
+      const onSave = () => {
+        if (_code !== code) {
+          departmentsEdit({ variables: { id: _id, code: _code } }, ['code']);
+        }
+      };
+
+      const onChange = (el: React.ChangeEvent<HTMLInputElement>) => {
+        setCode(el.currentTarget.value);
+      };
+
       return (
-        <RecordTablePopover>
-          <RecordTableCellTrigger>
-            {cell.getValue() as string}
-          </RecordTableCellTrigger>
-          <RecordTableCellContent>
-            <Input value={cell.getValue() as string} />
-          </RecordTableCellContent>
-        </RecordTablePopover>
+        <Popover
+          open={open}
+          onOpenChange={(open) => {
+            setOpen(open);
+            if (!open) onSave();
+          }}
+        >
+          <RecordTableInlineCell.Trigger>
+            <RecordTableTree.Trigger
+              order={(cell.row.original?.order as string) || ''}
+              name={cell.getValue() as string}
+              hasChildren={cell.row.original.hasChildren as boolean}
+            >
+              <TextOverflowTooltip value={cell.getValue() as string} />
+            </RecordTableTree.Trigger>
+          </RecordTableInlineCell.Trigger>
+          <RecordTableInlineCell.Content>
+            <Input value={_code} onChange={onChange} disabled={loading} />
+          </RecordTableInlineCell.Content>
+        </Popover>
       );
     },
   },
@@ -115,21 +152,39 @@ export const DepartmentColumns: ColumnDef<IDepartmentListItem>[] = [
     accessorKey: 'title',
     header: () => <RecordTable.InlineHead label="title" />,
     cell: ({ cell }) => {
+      const { departmentsEdit, loading } = useDepartmentInlineEdit();
+      const { _id, code, title } = cell.row.original;
+      const [_title, setTitle] = useState<string>(title);
+      const [open, setOpen] = useState<boolean>(false);
+
+      const onSave = () => {
+        if (_title !== title) {
+          departmentsEdit({ variables: { id: _id, title: _title, code } }, [
+            'title',
+            'code',
+          ]);
+        }
+      };
+
+      const onChange = (el: React.ChangeEvent<HTMLInputElement>) => {
+        setTitle(el.currentTarget.value);
+      };
+
       return (
-        <RecordTablePopover>
-          <RecordTableCellTrigger>
-            <RecordTableTree.Trigger
-              order={(cell.row.original?.order as string) || ''}
-              name={cell.getValue() as string}
-              hasChildren={cell.row.original?.hasChildren as boolean}
-            >
-              <TextOverflowTooltip value={cell.getValue() as string} />
-            </RecordTableTree.Trigger>
-          </RecordTableCellTrigger>
-          <RecordTableCellContent>
-            <Input value={cell.getValue() as string} />
-          </RecordTableCellContent>
-        </RecordTablePopover>
+        <Popover
+          open={open}
+          onOpenChange={(open) => {
+            setOpen(open);
+            if (!open) onSave();
+          }}
+        >
+          <RecordTableInlineCell.Trigger>
+            {cell.getValue() as string}
+          </RecordTableInlineCell.Trigger>
+          <RecordTableInlineCell.Content>
+            <Input value={_title} onChange={onChange} disabled={loading} />
+          </RecordTableInlineCell.Content>
+        </Popover>
       );
     },
     size: 350,
@@ -139,13 +194,24 @@ export const DepartmentColumns: ColumnDef<IDepartmentListItem>[] = [
     accessorKey: 'supervisorId',
     header: () => <RecordTable.InlineHead label="supervisor" />,
     cell: ({ cell }) => {
+      const { _id, code } = cell.row.original;
+      const { departmentsEdit } = useDepartmentInlineEdit();
       return (
-        <RecordTableCellDisplay>
-          <AssignMember
-            className="shadow-none bg-transparent"
+        <RecordTableInlineCell>
+          <SelectMember.InlineCell
+            scope={`DepartmentsPage.${_id}`}
             value={cell.getValue() as string}
+            mode="single"
+            onValueChange={(value) => {
+              departmentsEdit(
+                {
+                  variables: { id: _id, supervisorId: value, code },
+                },
+                ['supervisorId', 'code'],
+              );
+            }}
           />
-        </RecordTableCellDisplay>
+        </RecordTableInlineCell>
       );
     },
   },
@@ -155,9 +221,9 @@ export const DepartmentColumns: ColumnDef<IDepartmentListItem>[] = [
     header: () => <RecordTable.InlineHead label="team member count" />,
     cell: ({ cell }) => {
       return (
-        <RecordTableCellDisplay className="justify-center">
+        <RecordTableInlineCell className="justify-center">
           <Badge variant={'secondary'}>{cell.getValue() as number}</Badge>
-        </RecordTableCellDisplay>
+        </RecordTableInlineCell>
       );
     },
   },
@@ -166,11 +232,11 @@ export const DepartmentColumns: ColumnDef<IDepartmentListItem>[] = [
     header: () => <RecordTable.InlineHead label="Actions" />,
     cell: ({ cell }) => {
       return (
-        <RecordTableCellDisplay className="gap-1 [&>button]:px-2 justify-center">
+        <RecordTableInlineCell className="gap-1 [&>button]:px-2 justify-center">
           <DepartmentWorkingHoursColumnCell cell={cell} />
           <DepartmentMoreColumnCell cell={cell} />
           <DepartmentRemoveCell cell={cell} />
-        </RecordTableCellDisplay>
+        </RecordTableInlineCell>
       );
     },
   },
